face-recognition-faceid/requirements.txt
numpy
opencv-python
pillow
matplotlib
torch
torchvision
